argc: sema: seperate func
